refactor: moved simplifier logic to its own package
